chore: remove deprecated lime APIs
