+++ core/src/test/scala/com/microsoft/azure/synapse/ml/image/SuperpixelSuite.scala
@@ -1,9 +1,9 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lime
+package com.microsoft.azure.synapse.ml.image
 
-import com.microsoft.azure.synapse.ml.image.ImageTestUtils
+import com.microsoft.azure.synapse.ml.explainers.LIMEUtils
 import com.microsoft.azure.synapse.ml.io.image.ImageUtils
 
 import java.awt.Color
