+++ core/src/test/scala/com/microsoft/azure/synapse/ml/image/SuperpixelTransformerSuite.scala
@@ -1,10 +1,9 @@
 // Copyright (C) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in project root for information.
 
-package com.microsoft.azure.synapse.ml.lime
+package com.microsoft.azure.synapse.ml.image
 
 import com.microsoft.azure.synapse.ml.core.test.fuzzing.{TestObject, TransformerFuzzing}
-import com.microsoft.azure.synapse.ml.image.ImageTestUtils
 import com.microsoft.azure.synapse.ml.io.split1.FileReaderUtils
 import org.apache.spark.ml.util.MLReadable
